\name{is.POSIXt}
\alias{is.POSIXt}
\title{Test if an object is of class POSIXct or POSIXlt or POSIXt}
\usage{
is.POSIXt(x)
}
\arguments{
  \item{x}{Any R object}
}
\value{
Returns TRUE if object is of class POSIXct or POSIXlt or
POSIXt
}
\description{
Test if an object is of class POSIXct or POSIXlt or POSIXt
}
\author{
David M. Kaplan \email{dmkaplan2000@gmail.com}
}

