\docType{data}
\name{nxt.genesis.ts}
\alias{nxt.genesis.ts}
\title{Timestamp of NXT genesis block in POSIXct format}
\format{\preformatted{ POSIXct[1:1], format: "2013-11-24 12:00:00"
}}
\usage{
nxt.genesis.ts
}
\description{
Timestamp of NXT genesis block in POSIXct format
}
\author{
David M. Kaplan \email{dmkaplan2000@gmail.com}
}
\seealso{
\code{\link{nxt.convert.ts}}
}
\keyword{datasets}

