\name{RawToHexString}
\alias{RawToHexString}
\title{Converts raw R objects to hex string}
\usage{
RawToHexString(x)
}
\arguments{
  \item{x}{Object of type \code{\link{raw}}}
}
\value{
Hex string corresponding to raw object
}
\description{
Converts a \code{\link{raw}} R object to a hex strings
(e.g., "ac6fef")
}
\author{
David M. Kaplan \email{dmkaplan2000@gmail.com}
}
\seealso{
\code{\link{raw}}, \code{\link{rawToChar}},
\code{\link{HexStringToRaw}}
}

