\docType{package}
\name{Rxt}
\alias{Rxt}
\alias{Rxt-package}
\title{Functions for analyzing the H2 database containing the NXT blockchain}
\description{
Functions for analyzing the H2 database containing the NXT
blockchain
}
\details{
This package contains a series of R functions for accessing
and analyzing data in the NXT blockchain, which is stored
in a H2 (java) database.  To use this package, one must
either connect to a copy of the H2 database that is not in
active use by the (NXT) NRS Client or instruct the NRS
client to open the database with the AUTO_SERVER=TRUE
option using the nxt.dbUrl property.

\tabular{ll}{ Package: \tab Rxt\cr Type: \tab Package\cr
Version: \tab 0.2\cr Date: \tab 2014-04-02\cr License: \tab
GPL (>= 2)\cr }
}
\author{
David M. Kaplan \email{dmkaplan2000@gmail.com}

Maintainer: David M. Kaplan \email{dmkaplan2000@gmail.com}
}
\keyword{package}

